requirements.txt
django>=5.0,<5.1
gunicorn
django-extensions
IPython
psycopg2
environs
django-import-export
python-dateutil
tablib